src: main: java: csd: roster: service: DepartmentService: Update params for delete method
src: main: java: csd: roster: service: DepartmentServiceImpl: Add validation to check if company exist and lambda expression to delete department if it exists

+++ src/main/java/csd/roster/service/DepartmentServiceImpl.java
@@ -1,8 +1,12 @@
 package csd.roster.service;
 
+import csd.roster.exception.CompanyNotFoundException;
+import csd.roster.exception.DepartmentNotFoundException;
+import csd.roster.model.Company;
 import csd.roster.model.Department;
 import csd.roster.repository.DepartmentRepository;
 import org.hibernate.cfg.NotYetImplementedException;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,9 +17,11 @@ import java.util.UUID;
 public class DepartmentServiceImpl implements DepartmentService {
 
     private DepartmentRepository departmentRepository;
+    private CompanyService companyService;
 
-    public DepartmentServiceImpl(DepartmentRepository departmentRepository) {
+    public DepartmentServiceImpl(DepartmentRepository departmentRepository, CompanyService companyService) {
         this.departmentRepository = departmentRepository;
+        this.companyService = companyService;
     }
 
     @Override
@@ -44,7 +50,26 @@ public class DepartmentServiceImpl implements DepartmentService {
     }
 
     @Override
-    public void delete(Department department) {
+    public void delete(UUID companyId, UUID departmentId) {
+        if (companyService.getCompanyById(companyId) == null)
+            throw new CompanyNotFoundException(companyId);
+
+        Department department = getDepartmentByIdAndCompanyId(departmentId, companyId)
+                .orElseThrow(() -> new DepartmentNotFoundException(companyId, departmentId));
+
         departmentRepository.delete(department);
+    }
+
+    @Override
+    public Department update(UUID companyId, UUID departmentId, Department department) {
+        Company company = companyService.getCompanyById(companyId)
+                .orElseThrow(() -> new CompanyNotFoundException(companyId));
+
+        return getDepartmentByIdAndCompanyId(departmentId, companyId).map(oldDepartment -> {
+            department.setCompany(company);
+            department.setId(departmentId);
+            return departmentRepository.save(department);
+        }).orElseThrow(() -> new DepartmentNotFoundException(companyId, departmentId));
+
     }
 }
